refactor: reorganize toolbar directory into toolbar, viewport-controls, and alerts
The toolbar/ directory contained 21 files serving three unrelated purposes. Split into separate directories by concern: toolbar/ (12 files) for left-side toolbar buttons, viewport-controls/ (5 files) for viewport corner control strips, alerts/ (4 files) for root-level overlay/alert banners.

Made-with: Cursor
