+++ src/editor/alerts/alert-scene-limits.ts
